+++ test/mailers/business_mailer_test.rb
@@ -1,15 +1,15 @@
 require "test_helper"
 
 class BusinessMailerTest < ActionMailer::TestCase
-  test "welcome business email test" do
-    @business = Business.first
-    email = BusinessMailer.with(business: @business).welcome_email
+  test "sending the welcome email to a business" do
+    business = businesses(:one)
+    email = BusinessMailer.with(business: business).welcome_email
 
     assert_emails 1 do
       email.deliver_now
     end
 
-    assert_equal [@business.user.email], email.to
-    assert_match /added your business profile/, email.body.encoded
+    assert_equal [business.user.email], email.to
+    assert_match(/added your business/, email.body.encoded)
   end
 end